requirements.txt
Pillow==4.3.0
PyWavelets==0.5.2
argparse==1.2.1
backports.functools-lru-cache==1.4
certifi==2017.11.5
chainer==3.2.0
cycler==0.10.0
decorator==4.1.2
filelock==2.0.14
future==0.16.0
matplotlib==2.1.1
networkx==2.0
numpy==1.13.3
olefile==0.44
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-telegram-bot==9.0.0
pytz==2017.3
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
subprocess32==3.2.7
urllib3==1.22
wsgiref==0.1.2
